+++ migration-library/src/main/kotlin/com/quadient/migration/service/deploy/InteractiveDeployClient.kt
@@ -108,7 +108,7 @@ class InteractiveDeployClient(
             }
 
             try {
-                val documentObjectXml = documentObjectBuilder.buildDocumentObject(it)
+                val documentObjectXml = documentObjectBuilder.buildDocumentObject(it, null)
                 val runCommandType = it.type.toRunCommandType()
 
                 val editResult = ipsService.deployJld(