+++ migration-library/src/test/kotlin/com/quadient/migration/service/inspirebuilder/DesignerDocumentObjectBuilderTest.kt
@@ -34,6 +34,7 @@ import com.quadient.migration.shared.Position
 import com.quadient.migration.shared.VariablePathData
 import com.quadient.migration.shared.centimeters
 import com.quadient.migration.shared.millimeters
+import com.quadient.migration.shared.toIcmPath
 import com.quadient.migration.tools.aProjectConfig
 import com.quadient.migration.tools.model.aCell
 import com.quadient.migration.tools.model.aVariable
@@ -54,6 +55,7 @@ import io.mockk.verify
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 
@@ -113,7 +115,7 @@ class DesignerDocumentObjectBuilderTest {
 
         // when
         val result =
-            subject.buildDocumentObject(template).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
+            subject.buildDocumentObject(template, null).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
 
         // then
         result["Page"].size().shouldBeEqualTo(4)
@@ -171,7 +173,7 @@ class DesignerDocumentObjectBuilderTest {
 
         // when
         val result =
-            subject.buildDocumentObject(template).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
+            subject.buildDocumentObject(template, null).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
 
         // then
         result["Image"].shouldBeEqualTo(null)
@@ -204,7 +206,7 @@ class DesignerDocumentObjectBuilderTest {
 
         // when
         val result =
-            subject.buildDocumentObject(template).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
+            subject.buildDocumentObject(template, null).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
 
         // then
         val imageObject = result["ImageObject"].last()
@@ -236,7 +238,7 @@ class DesignerDocumentObjectBuilderTest {
 
         // when
         val result =
-            subject.buildDocumentObject(template).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
+            subject.buildDocumentObject(template, null).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
 
         // then
         val areaFlowId = result["FlowArea"].last()["FlowId"].textValue()
@@ -280,7 +282,7 @@ class DesignerDocumentObjectBuilderTest {
         )
 
         // when
-        val result = subject.buildDocumentObject(block).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
+        val result = subject.buildDocumentObject(block, null).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
 
         // then
         val rowSetId = result["Table"].last()["RowSetId"].textValue()
@@ -306,7 +308,7 @@ class DesignerDocumentObjectBuilderTest {
 
         // when
         val result =
-            subject.buildDocumentObject(template).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
+            subject.buildDocumentObject(template, null).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
 
         // then
         result["Flow"].filter { it["Name"]?.textValue() == block.nameOrId() }.size.shouldBeEqualTo(1)
@@ -331,7 +333,7 @@ class DesignerDocumentObjectBuilderTest {
 
         // when
         val result =
-            subject.buildDocumentObject(template).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
+            subject.buildDocumentObject(template, null).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
 
         // then
         val areaFlowId = result["FlowArea"].last()["FlowId"].textValue()
@@ -393,7 +395,7 @@ class DesignerDocumentObjectBuilderTest {
         // when
         val subject = aSubject(config)
         val result =
-            subject.buildDocumentObject(template).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
+            subject.buildDocumentObject(template, null).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
 
         // then
         val variableData = result["Variable"].single { it["Name"]?.textValue() == variable.nameOrId() }
@@ -448,7 +450,7 @@ class DesignerDocumentObjectBuilderTest {
         )
 
         // when
-        val result = subject.buildDocumentObject(block).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
+        val result = subject.buildDocumentObject(block, null).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
 
         // then
         val conditionFlow = result["Flow"].last { it["Type"]?.textValue() == "InlCond" }
@@ -496,7 +498,7 @@ class DesignerDocumentObjectBuilderTest {
             mockObj(aDocObj("B_1", Block, listOf(aParagraph(aText(StringModel("Text")))), displayRuleRef = rule.id))
 
         // when
-        val result = subject.buildDocumentObject(block).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
+        val result = subject.buildDocumentObject(block, null).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
 
         // then
         val flowAreaFlowId = result["FlowArea"].last()["FlowId"].textValue()
@@ -545,7 +547,7 @@ class DesignerDocumentObjectBuilderTest {
 
         // when
         val subject = aSubject(config)
-        val result = subject.buildDocumentObject(block).let { xmlMapper.readTree(it.trimIndent()) }
+        val result = subject.buildDocumentObject(block, null).let { xmlMapper.readTree(it.trimIndent()) }
 
         // then
         result["Property"].size().shouldBeEqualTo(3)
@@ -584,8 +586,8 @@ class DesignerDocumentObjectBuilderTest {
 
         // when
         val subject = aSubject(config)
-        subject.buildDocumentObject(block1)
-        subject.buildDocumentObject(block2)
+        subject.buildDocumentObject(block1, null)
+        subject.buildDocumentObject(block2, null)
 
         // then
         verify(exactly = 1) { ipsService.wfd2xml(any()) }
@@ -618,7 +620,7 @@ class DesignerDocumentObjectBuilderTest {
 
         // when
         val subject = aSubject(config)
-        val result = subject.buildDocumentObject(block).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
+        val result = subject.buildDocumentObject(block, null).let { xmlMapper.readTree(it.trimIndent()) }["Layout"]["Layout"]
 
         // then
         verify(exactly = 1) { variableStructureRepository.findModelOrFail(variableStructureA.id) }
@@ -736,22 +738,37 @@ class DesignerDocumentObjectBuilderTest {
 
         @ParameterizedTest
         @CsvSource(
-            // defaultTargetFolder  ,expected
-            "                       ,icm://projectNameStyles.wfd",
-            "null                   ,icm://projectNameStyles.wfd",
-            "default                ,icm://default/projectNameStyles.wfd",
+            // styleDefPath        ,defaultTargetFolder  ,expected
+            "                      ,                       ,icm://projectNameStyles.wfd",
+            "                      ,null                   ,icm://projectNameStyles.wfd",
+            "                      ,default                ,icm://default/projectNameStyles.wfd",
+            "                      ,default                ,icm://default/projectNameStyles.wfd",
+            "icm://some/path/f.wfd ,default                ,icm://some/path/f.wfd",
         )
-        fun testCompanyStylesPath(defaultTargetFolder: String?, expected: String) {
+        fun testCompanyStylesPath(styleDefPath: String?, defaultTargetFolder: String?, expected: String) {
             val config = aProjectConfig(
                 name = "projectName",
                 output = InspireOutput.Designer,
                 targetDefaultFolder = defaultTargetFolder.nullToNull(),
+                styleDefinitionPath = styleDefPath?.toIcmPath()
             )
             val pathTestSubject = aSubject(config)
 
             val path = pathTestSubject.getStyleDefinitionPath()
 
             path.shouldBeEqualTo(expected)
+        }
+
+        @Test
+        fun companyStylePathMustBeAbsolute() {
+            val config = aProjectConfig(
+                name = "projectName",
+                output = InspireOutput.Designer,
+                styleDefinitionPath = "somePath.wfd".toIcmPath()
+            )
+            val pathTestSubject = aSubject(config)
+
+            assertThrows<IllegalArgumentException> { pathTestSubject.getStyleDefinitionPath() }
         }
 
         private fun String?.nullToNull() = when (this?.trim()) {